golioth: move golioth_estimated_coap_block_size() to golioth_utils.c
Create new golioth_utils.c module, which will contain private utilities
shared among various golioth modules.
